move elasticlunr to it's own module
